refactor: extract server init code
This prepares for #1624, so the end results looks a bit cleaner.
