+++ src/influxdb_ioxd/http.rs
@@ -899,7 +899,7 @@ mod tests {
         let (_, config) = config();
         let app_server = Arc::new(AppServer::new(ConnectionManagerImpl {}, config));
         app_server.set_id(ServerId::try_from(1).unwrap()).unwrap();
-        app_server.maybe_initialize_server().await.unwrap();
+        app_server.maybe_initialize_server().await;
         app_server
             .create_database(DatabaseRules::new(
                 DatabaseName::new("MyOrg_MyBucket").unwrap(),
@@ -947,7 +947,7 @@ mod tests {
         let (metrics_registry, config) = config();
         let app_server = Arc::new(AppServer::new(ConnectionManagerImpl {}, config));
         app_server.set_id(ServerId::try_from(1).unwrap()).unwrap();
-        app_server.maybe_initialize_server().await.unwrap();
+        app_server.maybe_initialize_server().await;
         app_server
             .create_database(DatabaseRules::new(
                 DatabaseName::new("MetricsOrg_MetricsBucket").unwrap(),
@@ -1037,7 +1037,7 @@ mod tests {
         let (_, config) = config();
         let app_server = Arc::new(AppServer::new(ConnectionManagerImpl {}, config));
         app_server.set_id(ServerId::try_from(1).unwrap()).unwrap();
-        app_server.maybe_initialize_server().await.unwrap();
+        app_server.maybe_initialize_server().await;
         app_server
             .create_database(DatabaseRules::new(
                 DatabaseName::new("MyOrg_MyBucket").unwrap(),
@@ -1174,7 +1174,7 @@ mod tests {
         let (_, config) = config();
         let app_server = Arc::new(AppServer::new(ConnectionManagerImpl {}, config));
         app_server.set_id(ServerId::try_from(1).unwrap()).unwrap();
-        app_server.maybe_initialize_server().await.unwrap();
+        app_server.maybe_initialize_server().await;
         app_server
             .create_database(DatabaseRules::new(
                 DatabaseName::new("MyOrg_MyBucket").unwrap(),
@@ -1223,7 +1223,7 @@ mod tests {
         let (_, config) = config();
         let app_server = Arc::new(AppServer::new(ConnectionManagerImpl {}, config));
         app_server.set_id(ServerId::try_from(1).unwrap()).unwrap();
-        app_server.maybe_initialize_server().await.unwrap();
+        app_server.maybe_initialize_server().await;
         app_server
             .create_database(DatabaseRules::new(
                 DatabaseName::new("MyOrg_MyBucket").unwrap(),
